refactor: make server decomposition
